[refactoring] step3: 카드 generator 제거, 카드 덱 수정
