+++ src/main/kotlin/blackjack_dealer/domain/Dealer.kt
@@ -1,6 +1,5 @@
 package blackjack_dealer.domain
 
-import blackjack_dealer.CardGenerator
 import blackjack_dealer.entity.CardDeque
 import blackjack_dealer.entity.GamerCards
 import blackjack_dealer.entity.state.GamerCurrentState
