+++ src/main/kotlin/blackjack_dealer/domain/Participant.kt
@@ -1,10 +1,8 @@
 package blackjack_dealer.domain
 
-import blackjack_dealer.CardGenerator
 import blackjack_dealer.entity.CardDeque
 import blackjack_dealer.entity.GamerCards
 import blackjack_dealer.entity.state.GamerCurrentState
-import blackjack_dealer.entity.toGamerCards
 
 data class Participant(
     override val name: String,
