+++ src/main/kotlin/main.kt
@@ -1,6 +1,5 @@
 import blackjack_dealer.BlackJack
 import blackjack_dealer.BlackJackResultBoard
-import blackjack_dealer.CardGenerator
 import blackjack_dealer.domain.Dealer
 import blackjack_dealer.entity.BlackJackGamer
 import blackjack_dealer.entity.CardDeque
@@ -13,8 +12,8 @@ fun main() {
     val participantsName = InputView.inputParticipantsName()
     val cardDeque = CardDeque.create()
 
-    val participants = Participants.newInstance(participantsName) { CardGenerator.generateDoubleCard(cardDeque) }
-    val dealer = Dealer.newInstance(CardGenerator.generateDoubleCard(cardDeque))
+    val participants = Participants.newInstance(participantsName) { cardDeque.generateDoubleCard() }
+    val dealer = Dealer.newInstance(cardDeque.generateDoubleCard())
     OutputView.printDivideCardsToGamer(dealer, participants)
 
     OutputView.printGamersInformation(dealer, participants)
