--- src/test/kotlin/blackjack_dealer/CardGeneratorTest.kt
@@ -1,21 +1,0 @@
-package blackjack_dealer
-
-import blackjack_dealer.entity.CardDeque
-import io.kotest.core.spec.style.StringSpec
-import io.kotest.matchers.shouldBe
-
-class CardGeneratorTest : StringSpec({
-    "CardGenerator 의 한장 생성 함수가 잘 작동 한다" {
-        val cardDeque = CardDeque.create()
-        val expected = 51
-        CardGenerator.generateSingleCard(cardDeque)
-        cardDeque.cardDeque.count() shouldBe expected
-    }
-
-    "CardGenerator 의 두장 생성 함수가 잘 작동한다" {
-        val cardDeque = CardDeque.create()
-        val expected = 50
-        CardGenerator.generateDoubleCard(cardDeque)
-        cardDeque.cardDeque.count() shouldBe expected
-    }
-})
